requirements.txt
pandas
gspread
boto3
psycopg2-binary
pyarrow==23.0.1
s3fs 
tenacity
google-api-python-client



